+++ inst/htmlwidgets/RacViewer/lib/functions/agsr.js
@@ -101,6 +101,7 @@ Racmacs.Point = class Point {
         this.rayTraceable = true;
         this.opacity      = 1;
         this.scaling      = 1;
+        this.included     = true;
 
 
         this.coords_na = args.coords === null || isNaN(args.coords[0]) || args.coords[0] === null;
@@ -266,7 +267,7 @@ Racmacs.Point = class Point {
     }
 
     // Set the point position
-    setPosition(to){
+    setPosition(to, dispatch = true){
 
         while (to.length < 3) to.push(0);
 
@@ -303,11 +304,13 @@ Racmacs.Point = class Point {
         }
 
         // Update any connections or error lines
-        this.viewer.dispatchEvent("point-moved", {
-            point : this,
-            from  : from,
-            to    : to
-        });
+        if (dispatch) {
+            this.viewer.dispatchEvent("point-moved", {
+                point : this,
+                from  : from,
+                to    : to
+            });
+        }
 
     }
 
@@ -621,6 +624,56 @@ Racmacs.Point = class Point {
         return(stress);
     }
 
+    // Toggle whether the point is included or not
+    toggleIncluded() {
+
+        if (this.included) {
+            this.excludePoint();
+        } else {
+            this.includePoint();
+        }
+
+    }
+
+    // Include a point
+    includePoint() {
+
+        if (!this.included) {
+
+            this.viewer.data.includePoint(this.type, this.typeIndex);
+            this.included = true;
+            this.shown = true;
+            this.element.show();
+            this.viewer.render();
+            this.viewer.dispatchEvent("point-included", { 
+                point : this 
+            });
+            this.viewer.updateStress();
+
+        }
+
+    }
+
+    // Exclude a point
+    excludePoint() {
+
+        if (!this.excluded) {
+
+            this.viewer.data.excludePoint(this.type, this.typeIndex);
+            this.included = false;
+            this.shown = false;
+            this.element.hide();
+            this.viewer.render();
+            this.viewer.dispatchEvent("point-excluded", { 
+                point : this 
+            });
+            this.viewer.updateStress();
+
+        }
+
+    }
+
+
 }
 
 
@@ -636,6 +689,9 @@ Racmacs.Antigen = class Antigen extends Racmacs.Point {
     }
 
     // Get the titers to another point
+    titers(){
+        return(this.viewer.data.titertable[this.typeIndex]);
+    }
     titerTo(to){ 
         return(this.viewer.data.titertable[this.typeIndex][to.typeIndex]);
     }
@@ -661,6 +717,9 @@ Racmacs.Serum = class Serum extends Racmacs.Point {
     }
 
     // Get the titers to another point
+    titers(){
+        return(this.viewer.data.titertable.map(row => row[this.typeIndex]));
+    }
     titerTo(to){ 
         return(this.viewer.data.titertable[to.typeIndex][this.typeIndex]);
     }